+++ MediaExtractor/MainWindow.xaml.cs
@@ -887,6 +887,16 @@ namespace MediaExtractor
         }
 
         /// <summary>
+        /// Enables Italian as application language
+        /// </summary>
+        /// <param name="sender">The source of the event.</param>
+        /// <param name="e">The <see cref="RoutedEventArgs"/> instance containing the event data.</param>
+        private void ItalianMenuItem_Click(object sender, RoutedEventArgs e)
+        {
+            ChangeLocale(I18n.ITALIAN);
+        }
+
+        /// <summary>
         /// Enables the system locale as application language
         /// </summary>
         /// <param name="sender">The source of the event.</param>