+++ SetupProject/Constants.cs
@@ -55,6 +55,7 @@ namespace WixSharp
         public const string LANGUAGE_GERMAN = "DE";
         public const string LANGUAGE_SPANISH = "ES";
         public const string LANGUAGE_JAPANESE = "JP";
+        public const string LANGUAGE_ITALIAN = "IT";
 
         public const string INSTALLATION_TYPE_USER = "user";
         public const string INSTALLATION_TYPE_SYSTEM = "system";